Added Permission based Functionality

+++ academics/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import get_object_or_404, render, redirect
+from django.shortcuts import get_object_or_404, render
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 from .models import Department, Program, Semester, Course, Batch
@@ -7,10 +7,14 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 def department(request):
-    departments = Department.objects.all()
-    form = DepartmentForm()
+    if request.user.role == 'Admin':
+        departments = Department.objects.all()
+    elif request.user.role in ['Faculty', 'Editor']:
+        departments = Department.objects.filter(department_id=request.user.department.department_id)
+    else:
+        departments = Department.objects.none()
 
-    # Create a dictionary of forms for each department
+    form = DepartmentForm()
     update_forms = {department.department_id: DepartmentForm(instance=department) for department in departments}
 
     return render(request, 'academics/departments.html', {
@@ -35,6 +39,10 @@ def add_department(request):
 @login_required
 def department_update(request, department_id):
     department = get_object_or_404(Department, department_id=department_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to update departments.'})
+
     if request.method == 'POST':
         form = DepartmentForm(request.POST, instance=department)
         if form.is_valid():
@@ -48,6 +56,10 @@ def department_update(request, department_id):
 @login_required
 def department_delete(request, department_id):
     department = get_object_or_404(Department, department_id=department_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to delete departments.'})
+
     if request.method == 'POST':
         department.delete()
         return JsonResponse({'success': True, 'message': f'Department {department.department_name} deleted successfully.'})
@@ -56,13 +68,17 @@ def department_delete(request, department_id):
 # Program Section 
 @login_required
 def program(request):
-    programs = Program.objects.all()
-    form = ProgramForm()
+    if request.user.role == 'Admin':
+        programs = Program.objects.all()
+    elif request.user.role in ['Faculty', 'Editor']:
+        programs = Program.objects.filter(department_id=request.user.department.department_id)
+    else:
+        programs = Program.objects.none()
 
-    # Create a dictionary of forms for each program
+    form = ProgramForm()
     update_forms = {program.program_id: ProgramForm(instance=program) for program in programs}
 
-    return render(request, 'academics\program.html', {
+    return render(request, 'academics/program.html', {
         'programs': programs,
         'form': form,
         'update_forms': update_forms,
@@ -84,6 +100,10 @@ def add_program(request):
 @login_required
 def program_update(request, program_id):
     program = get_object_or_404(Program, program_id=program_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to update programs.'})
+
     if request.method == 'POST':
         form = ProgramForm(request.POST, instance=program)
         if form.is_valid():
@@ -97,21 +117,32 @@ def program_update(request, program_id):
 @login_required
 def program_delete(request, program_id):
     program = get_object_or_404(Program, program_id=program_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to delete programs.'})
+
     if request.method == 'POST':
         program.delete()
-        return JsonResponse({'success': True, 'message': f'Department {program.program_name} deleted successfully.'})
+        return JsonResponse({'success': True, 'message': f'Program {program.program_name} deleted successfully.'})
     return JsonResponse({'success': False, 'message': 'Invalid request method'})
 
 # Semester Section
 @login_required
 def semester(request):
-    semesters = Semester.objects.all()
-    form = SemesterForm()
+    if request.user.role == 'Admin':
+        semesters = Semester.objects.all()
+    elif request.user.role in ['Faculty', 'Editor']:
+        # Filter based on the program's department ID
+        departments = Department.objects.filter(department_id=request.user.department.department_id)
+        programs = Program.objects.filter(department__in=departments)
+        semesters = Semester.objects.filter(program__in=programs)
+    else:
+        semesters = Semester.objects.none()
 
-    # Create a dictionary of forms for each department
+    form = SemesterForm()
     update_forms = {semester.semester_id: SemesterForm(instance=semester) for semester in semesters}
 
-    return render(request, 'academics\semester.html', {
+    return render(request, 'academics/semester.html', {
         'semesters': semesters,
         'form': form,
         'update_forms': update_forms,
@@ -133,6 +164,10 @@ def add_semester(request):
 @login_required
 def semester_update(request, semester_id):
     semester = get_object_or_404(Semester, semester_id=semester_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to update semesters.'})
+
     if request.method == 'POST':
         form = SemesterForm(request.POST, instance=semester)
         if form.is_valid():
@@ -146,6 +181,10 @@ def semester_update(request, semester_id):
 @login_required
 def semester_delete(request, semester_id):
     semester = get_object_or_404(Semester, semester_id=semester_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to delete semesters.'})
+
     if request.method == 'POST':
         semester.delete()
         return JsonResponse({'success': True, 'message': f'Semester {semester.semester_number} deleted successfully.'})
@@ -154,13 +193,20 @@ def semester_delete(request, semester_id):
 # Course Section 
 @login_required
 def course(request):
-    courses = Course.objects.all()
-    form = CourseForm()
+    if request.user.role == 'Admin':
+        courses = Course.objects.all()
+    elif request.user.role in ['Faculty', 'Editor']:
+        # Filter based on the department ID
+        departments = Department.objects.filter(department_id=request.user.department.department_id)
+        programs = Program.objects.filter(department__in=departments)
+        courses = Course.objects.filter(semester__program__in=programs)
+    else:
+        courses = Course.objects.none()
 
-    # Create a dictionary of forms for each department
+    form = CourseForm()
     update_forms = {course.course_id: CourseForm(instance=course) for course in courses}
 
-    return render(request, 'academics\course.html', {
+    return render(request, 'academics/course.html', {
         'courses': courses,
         'form': form,
         'update_forms': update_forms,
@@ -182,6 +228,10 @@ def add_course(request):
 @login_required
 def course_update(request, course_id):
     course = get_object_or_404(Course, course_id=course_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to update courses.'})
+
     if request.method == 'POST':
         form = CourseForm(request.POST, instance=course)
         if form.is_valid():
@@ -195,21 +245,32 @@ def course_update(request, course_id):
 @login_required
 def course_delete(request, course_id):
     course = get_object_or_404(Course, course_id=course_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to delete courses.'})
+
     if request.method == 'POST':
         course.delete()
         return JsonResponse({'success': True, 'message': f'Course {course.course_name} deleted successfully.'})
     return JsonResponse({'success': False, 'message': 'Invalid request method'})
 
-# Batch Section
+# Batch Section 
 @login_required
 def batch(request):
-    batches = Batch.objects.all()
-    form = BatchForm()
+    if request.user.role == 'Admin':
+        batches = Batch.objects.all()
+    elif request.user.role in ['Faculty', 'Editor']:
+        # Filter based on the department ID
+        departments = Department.objects.filter(department_id=request.user.department.department_id)
+        programs = Program.objects.filter(department__in=departments)
+        batches = Batch.objects.filter(program__in=programs)
+    else:
+        batches = Batch.objects.none()
 
-    # Create a dictionary of forms for each department
+    form = BatchForm()
     update_forms = {batch.batch_id: BatchForm(instance=batch) for batch in batches}
 
-    return render(request, 'academics\programs_batches.html', {
+    return render(request, 'academics/programs_batches.html', {
         'batches': batches,
         'form': form,
         'update_forms': update_forms,
@@ -231,6 +292,10 @@ def add_batch(request):
 @login_required
 def batch_update(request, batch_id):
     batch = get_object_or_404(Batch, batch_id=batch_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to update batches.'})
+
     if request.method == 'POST':
         form = BatchForm(request.POST, instance=batch)
         if form.is_valid():
@@ -244,6 +309,10 @@ def batch_update(request, batch_id):
 @login_required
 def batch_delete(request, batch_id):
     batch = get_object_or_404(Batch, batch_id=batch_id)
+    
+    if request.user.role != 'Admin':
+        return JsonResponse({'success': False, 'message': 'You do not have permission to delete batches.'})
+
     if request.method == 'POST':
         batch.delete()
         return JsonResponse({'success': True, 'message': f'Batch {batch.batch_name} deleted successfully.'})
